유휴 시간 보상 팝업 추가 및 자료형 변경
유휴 시간에 따른 보상 팝업 추가하고 자료형 Double로 변경

+++ 2D Clicker Demonstration/Assets/Scripts/Purchasables/CharacterButton.cs
@@ -23,7 +23,7 @@ namespace Purchasables
         public int initialCurrentCost = 1;    
 
         [HideInInspector]
-        public int goldPerSec;
+        public double goldPerSec;
         public int initialGoldPerSec = 1;
 
         public float upgradePower = 1.05f;
